+++ common_src/coordinates.h
@@ -9,16 +9,16 @@ accedidos con tanta frecuencia que conviene hacerlos
 publicos en lugar de llenar mi código de getters y setters, capaz me conviene convertir esto en un struct*/
 class Coordinates{
 public:
-    float x;
-    float y;
+    double x;
+    double y;
 
     Coordinates();
 
-    Coordinates(float a_x, float a_y);
+    Coordinates(double a_x, double a_y);
 
-    void increment_on_direction(Coordinates movement_direction, float inc);
+    void increment_on_direction(Coordinates movement_direction, double inc);
 
-    void rotate(float inc);
+    void rotate(double inc);
 
     void normalice_direction();
 
@@ -33,30 +33,30 @@ public:
 
     bool y_whole()const;
 
-    void inc_x(float a_x);
+    void inc_x(double a_x);
 
-    void inc_y(float a_y);
+    void inc_y(double a_y);
 
     bool has_element(const std::vector<std::vector<int>> &map);
 
-    float get_distance_to_lower_side_x()const;
+    double get_distance_to_lower_side_x()const;
 
-    float get_distance_to_higher_side_x()const;
+    double get_distance_to_higher_side_x()const;
 
-    float get_distance_to_lower_side_y()const;
+    double get_distance_to_lower_side_y()const;
 
-    float get_distance_to_higher_side_y()const;
+    double get_distance_to_higher_side_y()const;
 
-    float calculate_distance(const Coordinates &a_position) const;
+    double calculate_distance(const Coordinates &a_position) const;
 
 	// Según la direccion a la que apunta mi rayo, me fijo en una celda distinta
     int get_increase_x()const;
 
     int get_increase_y()const;
 
-    // void set_x(float a_x);
+    // void set_x(double a_x);
 
-    // void set_y(float a_y);
+    // void set_y(double a_y);
 
     //Compara nomas la parte "entera" de las coordenadas
     bool operator==(const Coordinates& other) const{
